Fixes for pytest
- reorganize tests
- delete service directory
- add true assertion for isupper test
