--- pytest_practice/0_list_tests/conftest.py
@@ -1,13 +1,0 @@
-import pytest
-
-
-@pytest.fixture()
-def lst():
-    return [0, 5, 10, 15, 20]
-
-
-@pytest.fixture()
-def lst1():
-    return [100, 200, 300, 400]
-
-
