--- pytest_practice/0_list_tests/test_list_1.py
@@ -1,3 +1,0 @@
-def test_insert_element(lst):
-    lst.insert(5, 25)
-    assert 25 in lst
