--- pytest_practice/0_list_tests/test_list_2.py
@@ -1,3 +1,0 @@
-def test_remove_element(lst):
-    del lst[0]
-    assert 0 not in lst
