--- pytest_practice/0_list_tests/test_list_3.py
@@ -1,3 +1,0 @@
-def test_extend_list(lst, lst1):
-    lst.extend(lst1)
-    assert len(lst) == 9
