--- pytest_practice/0_list_tests/test_list_4.py
@@ -1,3 +1,0 @@
-def test_reverse_list(lst):
-    lst.reverse()
-    assert lst[0] == 20
