--- pytest_practice/1_set_tests/conftest.py
@@ -1,13 +1,0 @@
-import pytest
-
-
-@pytest.fixture()
-def st():
-    return {'zero', 1, 'two', 3, 'four'}
-
-
-@pytest.fixture()
-def st1():
-    return {5, 6, 7, 'zero'}
-
-
