--- pytest_practice/1_set_tests/test_set_1.py
@@ -1,4 +1,0 @@
-def test_sets_int(st, st1):
-    assert len(st.union(st1)) == 8
-
-
