--- pytest_practice/1_set_tests/test_set_2.py
@@ -1,2 +1,0 @@
-def test_sets_union(st, st1):
-    assert len(st.intersection(st1)) == 1
