--- pytest_practice/1_set_tests/test_set_3.py
@@ -1,2 +1,0 @@
-def test_sets_diff(st, st1):
-    assert 'zero' not in st.difference(st1)
