--- pytest_practice/1_set_tests/test_set_4.py
@@ -1,2 +1,0 @@
-def test_sets_idj(st, st1):
-    assert st.isdisjoint(st1) is False
