--- pytest_practice/2_dictionary_tests/conftest.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-@pytest.fixture()
-def dct():
-    return {'John': 20, 'Kate': 23, 'Sam': 35, "Will": 19}
