--- pytest_practice/2_dictionary_tests/test_dictionary_1.py
@@ -1,3 +1,0 @@
-def test_dict_values(dct):
-    assert dct['John'] == 20
-    assert dct['Will'] == 19
