--- pytest_practice/2_dictionary_tests/test_dictionary_2.py
@@ -1,3 +1,0 @@
-def test_clear_dict(dct):
-    dct.clear()
-    assert len(dct) == 0
