--- pytest_practice/2_dictionary_tests/test_dictionary_3.py
@@ -1,3 +1,0 @@
-def test_del_dict(dct):
-    del dct['Will']
-    assert 'Will' not in dct
