--- pytest_practice/2_dictionary_tests/test_dictionary_4.py
@@ -1,3 +1,0 @@
-def test_set_dict(dct):
-    dct['Kate'] = 25
-    assert dct['Kate'] == 25
