--- pytest_practice/3_string_tests/test_string_1.py
@@ -1,2 +1,0 @@
-def test_str_find(stg):
-    assert stg.find('Строка') == 0
