--- pytest_practice/3_string_tests/test_string_2.py
@@ -1,2 +1,0 @@
-def test_str_islower(stg):
-    assert stg.islower() is False
