--- pytest_practice/3_string_tests/test_string_3.py
@@ -1,2 +1,0 @@
-def test_str_isspace(stg):
-    assert stg.isspace() is False
