--- pytest_practice/3_string_tests/test_string_4.py
@@ -1,4 +1,0 @@
-def test_str_add(stg, stg1):
-    assert len(stg+stg1) == 87
-
-
