--- pytest_practice/3_string_tests/test_string_5.py
@@ -1,9 +1,0 @@
-import pytest
-
-
-@pytest.mark.parametrize("up", [
-    'ЭТО ВЕРХНИЙ РЕГИСТР',
-    'а это не верхний регистр'
-])
-def test_is_upper(up):
-    assert up.isupper()
